feat(core): expose `detailedTypeof` and `isPlainObject` from core package
