+++ packages/core/src/index.ts
@@ -2,7 +2,9 @@ export * from './api/constants'
 export * from './api/actionFactory'
 export * from './api/createReducer'
 export * from './api/createStore'
+export * from './api/detailedTypeof'
+export * from './api/is'
+export * from './api/isPlainObject'
 export * from './api/MachineStore'
 export * from './api/meta'
-export * from './api/is'
 export * from './types'
